pyproject.toml
[tool.poetry]
name = "pyzetasql"
version = "0.1.0"
description = ""
authors = ["Katsuya Shimabukuro <katsu.generation.888@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.8"
grpcio = "^1.43.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
